services/chunker/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.74.0
langchain-text-splitters==0.3.4
structlog==24.4.0
tiktoken==0.8.0
